added compression for session state; made dbcontext disposable

+++ Serverless.Forum/Serverless.Forum/forum/PhpbbTopicsTrack.cs
@@ -8,6 +8,6 @@ namespace Serverless.Forum.forum
         public int UserId { get; set; }
         public int TopicId { get; set; }
         public int ForumId { get; set; }
-        public int MarkTime { get; set; }
+        public long MarkTime { get; set; }
     }
 }
